requirements.txt
black==22.3.0
jira==3.2.0
python-dotenv==0.20.0
ruamel.yaml==0.17.21
bokeh==2.4.3
selenium==4.2.0
selenium-chromedriver==1.3
PyPDF2==2.2.0
